Prevent authors from submitting essays in Studio Preview mode
